[로또 - 2단계] refactor 패키지 이름 변경
