+++ src/main/java/lotto/domain/AutoLottoMatchCount.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package lotto.domain;
 
 import java.util.HashMap;
 import java.util.Map;
