+++ src/main/java/lotto/util/PrizeCalculatorUtil.java
@@ -1,4 +1,4 @@
-package step2.util;
+package lotto.util;
 
 public class PrizeCalculatorUtil {
     private PrizeCalculatorUtil() {
