+++ src/test/java/lotto/UserTest.java
@@ -1,9 +1,8 @@
-package step2;
+package lotto;
 
 import org.junit.jupiter.api.Test;
-import step2.domain.AutoLottoMatchCount;
-import step2.domain.Lotto;
-import step2.domain.User;
+import lotto.domain.AutoLottoMatchCount;
+import lotto.domain.Lotto;
 
 import java.util.Arrays;
 import java.util.List;
