+++ src/test/java/step1/StringAddCalculatorTest.java
@@ -3,7 +3,7 @@ package step1;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import step1.domain.StringAddCalculator;
+import stringAddCalculator.domain.StringAddCalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
